feat(specs): updated structure for actions.
