--- src/actions/clear.action.ts
@@ -1,5 +1,0 @@
-import { dbConnector } from "../db";
-
-export function clearTasks(): Promise<void> {
-    return dbConnector.clearTasks();
-  }
