+++ src/actions/clear/clear.action.ts
@@ -1,0 +1,5 @@
+import {dbConnector} from "../../db";
+
+export function clearTasks(): Promise<void> {
+  return dbConnector.clearTasks();
+}
