+++ src/actions/index.ts
@@ -1,8 +1,8 @@
-import {add} from "./add.action";
-import {clearTasks} from "./clear.action";
-import {deleteTask} from "./delete.action";
-import { list } from "./list.action";
-import {update} from "./update.action";
+import {add} from "./add/add.action";
+import {clearTasks} from "./clear/clear.action";
+import {deleteTask} from "./delete/delete.action";
+import {list} from "./list/list.action";
+import {update} from "./update/update.action";
 
 export type ACTION_NAMES = "add" | "get" | "delete" | "list";
